new link handler added
