+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod handlers;
 pub mod link_handler;
+pub mod new_handler;
 pub mod router;
 pub mod server;
 pub mod thread_pool;
